test CI

+++ scripts/cpython-build-emsdk.sh
@@ -7,8 +7,7 @@
 # for allowing to avoid pyc creation
 
 
-
- PYTHON_FOR_BUILD=${PYTHON_FOR_BUILD:-${HOST_PREFIX}/bin/python3}
+export PYTHON_FOR_BUILD=${PYTHON_FOR_BUILD:-${HOST_PREFIX}/bin/python3}
 
 if [ -d $EMSDK ]
 then
@@ -28,7 +27,6 @@ then
 fi
 
 
-
 if [ -f build/cpython-wasm/libpython3.??.a ]
 then
     echo "
@@ -39,25 +37,31 @@ else
     * rebuilding build/cpython-wasm for [$PYDK_PYTHON_HOST_PLATFORM]
         PYTHON_FOR_BUILD=${PYTHON_FOR_BUILD}
 "
-    mkdir -p build/libffi $PREFIX
-    pushd build/libffi
 
-#TODO: check if export PATH=${HOST_PREFIX}/bin:$PATH is really set to avoid system python with different bytecode
-#and no loder lib-dynload in the way.
+    if [ -f ${PREFIX}/lib/libffi.a ]
+    then
+        echo "
+    * ffi already built"
+    else
+        mkdir -p build/libffi $PREFIX
+        pushd build/libffi
 
-    export EMCC_CFLAGS="-Os -g0 -fPIC"
+    #TODO: check if export PATH=${HOST_PREFIX}/bin:$PATH is really set to avoid system python with different bytecode
+    #and no loder lib-dynload in the way.
 
-     CFLAGS="-Os -g0 -fPIC"\
-      emconfigure $ROOT/src/libffi/configure --host=wasm32-tot-linux\
-      --prefix=$PREFIX --enable-static --disable-shared --disable-dependency-tracking\
-      --disable-builddir --disable-multi-os-directory --disable-raw-api --disable-docs\
+        export EMCC_CFLAGS="-Os -g0 -fPIC"
 
-    emmake make install
-    popd
+         CFLAGS="-Os -g0 -fPIC"\
+          emconfigure $ROOT/src/libffi/configure --host=wasm32-tot-linux\
+          --prefix=$PREFIX --enable-static --disable-shared --disable-dependency-tracking\
+          --disable-builddir --disable-multi-os-directory --disable-raw-api --disable-docs\
 
-    cp -fv  ${PREFIX}/lib/libffi.a $EMSDK/upstream/emscripten/cache/sysroot/lib/wasm32-emscripten/pic/
-    cp -vf  ${PREFIX}/include/ffi*.h ${EMSDK}/upstream/emscripten/cache/sysroot/include/
+        emmake make install
+        popd
 
+        cp -fv  ${PREFIX}/lib/libffi.a $EMSDK/upstream/emscripten/cache/sysroot/lib/wasm32-emscripten/pic/
+        cp -vf  ${PREFIX}/include/ffi*.h ${EMSDK}/upstream/emscripten/cache/sysroot/include/
+    fi
 
 
     mkdir -p build/cpython-wasm $PREFIX
